test-rmem-transfer: Change allocation of source/destination memory to kmalloc()
The memory allocation for source/destination was allocated from
reserved-memory. But from the point of view of the driver framework,
it's desiable for the driver to handle the kernel memory.

Change the source/destination memory to allocate from the kernel memory
with kmalloc(), and map the memory with dma_map_single() to DMA-space for
the device.

Fixed-memory is specified from the first property of memory-region in the
devicetree.

Signed-off-by: Kunihiko Hayashi <hayashi.kunihiko@socionext.com>